Move the compositor-side backdrop blur out of WaylandState
`ext-background-effect-v1` is the half of `backdrop_blur` the client cannot
do for itself — blurring what sits behind a translucent surface, which only
the compositor can see. Two fields and the request that pushes a region,
into `platform::backdrop`.

The rule worth keeping next to the code: a surface that never asked for blur
is left alone, because declaring an empty region overrides compositor-side
rules such as blur-by-namespace. Once it has blurred, dropping to zero rects
sends an empty region and never NULL — NULL withdraws our opinion and lets
those rules blur everything, where an empty region says "blur nothing".

wayland.rs 897 -> 787 lines, 17 -> 16 fields.
